dtslint
Tested with dtslint
